requirements.txt
Flask==3.0.3
flask-cors==5.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
httpx==0.27.0
pandas==2.3.3

langchain-groq==0.1.9
langchain-core==0.2.28
langchain-community==0.2.10
groq==0.4.1

schedule==1.2.0

sentinelhub==3.11.3
matplotlib==3.10.7
numpy>=1.26.4,<2.0.0
shapely>=2.0.0

earthengine-api==1.7.10